Safari-611.1.10/LayoutTests/http/tests/loading/promote-img-preload-priority.html

<html>
    <head>
        <script>
            if (window.testRunner) {
                testRunner.dumpResourceRequestPriorities();
                testRunner.dumpAsText();
            }
        </script>
        <script src="/resources/js-test-pre.js"></script>
    </head>
    <body>
        <img src="../misc/resources/image-slow.pl">
        <script src="/resources/js-test-post.js"></script>
    </body>
</html>
